requirements.txt
asgiref==3.2.10
attrs==19.3.0
Django==3.0.8
fuzzysearch==0.7.3
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
